+++ packages/apollo-server-core/src/ApolloServer.ts
@@ -16,7 +16,10 @@ import {
   DocumentNode,
 } from 'graphql';
 import { GraphQLExtension } from 'graphql-extensions';
-import { InMemoryLRUCache } from 'apollo-server-caching';
+import {
+  InMemoryLRUCache,
+  PrefixingKeyValueCache,
+} from 'apollo-server-caching';
 import { ApolloServerPlugin } from 'apollo-server-plugin-base';
 import runtimeSupportsUploads from './utils/runtimeSupportsUploads';
 
@@ -40,8 +43,6 @@ import {
   PluginDefinition,
 } from './types';
 
-import { FormatErrorExtension } from './formatters';
-
 import { gql } from './index';
 
 import {
@@ -50,10 +51,12 @@ import {
 } from './playground';
 
 import { generateSchemaHash } from './utils/schemaHash';
+import createSHA from './utils/createSHA';
 import {
   processGraphQLRequest,
   GraphQLRequestContext,
   GraphQLRequest,
+  APQ_CACHE_PREFIX,
 } from './requestPipeline';
 
 import { Headers } from 'apollo-server-env';
@@ -72,20 +75,18 @@ const NoIntrospection = (context: ValidationContext) => ({
   },
 });
 
-function getEngineServiceId(engine: Config['engine']): string | undefined {
+function getEngineApiKey(engine: Config['engine']): string | undefined {
   const keyFromEnv = process.env.ENGINE_API_KEY || '';
-  if (!(engine || (engine !== false && keyFromEnv))) {
-    return;
-  }
-
-  let engineApiKey: string = '';
-
   if (typeof engine === 'object' && engine.apiKey) {
-    engineApiKey = engine.apiKey;
+    return engine.apiKey;
   } else if (keyFromEnv) {
-    engineApiKey = keyFromEnv;
+    return keyFromEnv;
   }
+  return;
+}
 
+function getEngineServiceId(engine: Config['engine']): string | undefined {
+  const engineApiKey = getEngineApiKey(engine);
   if (engineApiKey) {
     return engineApiKey.split(':', 2)[1];
   }
@@ -108,6 +109,7 @@ export class ApolloServerBase {
   private context?: Context | ContextFunction;
   private engineReportingAgent?: import('apollo-engine-reporting').EngineReportingAgent;
   private engineServiceId?: string;
+  private engineApiKeyHash?: string;
   private extensions: Array<() => GraphQLExtension>;
   private schemaHash: string;
   protected plugins: ApolloServerPlugin[] = [];
@@ -209,11 +211,14 @@ export class ApolloServerBase {
     }
 
     if (requestOptions.persistedQueries !== false) {
-      if (!requestOptions.persistedQueries) {
-        requestOptions.persistedQueries = {
-          cache: requestOptions.cache!,
-        };
-      }
+      requestOptions.persistedQueries = {
+        cache: new PrefixingKeyValueCache(
+          (requestOptions.persistedQueries &&
+            requestOptions.persistedQueries.cache) ||
+            requestOptions.cache!,
+          APQ_CACHE_PREFIX,
+        ),
+      };
     } else {
       // the user does not want to use persisted queries, so we remove the field
       delete requestOptions.persistedQueries;
@@ -324,22 +329,18 @@ export class ApolloServerBase {
     // or cacheControl.
     this.extensions = [];
 
-    const debugDefault =
-      process.env.NODE_ENV !== 'production' && process.env.NODE_ENV !== 'test';
-    const debug =
-      requestOptions.debug !== undefined ? requestOptions.debug : debugDefault;
-
-    // Error formatting should happen after the engine reporting agent, so that
-    // engine gets the unmasked errors if necessary
-    this.extensions.push(
-      () => new FormatErrorExtension(requestOptions.formatError, debug),
-    );
-
     // In an effort to avoid over-exposing the API key itself, extract the
     // service ID from the API key for plugins which only needs service ID.
     // The truthyness of this value can also be used in other forks of logic
     // related to Engine, as is the case with EngineReportingAgent just below.
     this.engineServiceId = getEngineServiceId(engine);
+
+    const apiKey = getEngineApiKey(engine);
+    if (apiKey) {
+      this.engineApiKeyHash = createSHA('sha256')
+        .update(apiKey)
+        .digest('hex');
+    }
 
     if (this.engineServiceId) {
       const { EngineReportingAgent } = require('apollo-engine-reporting');
@@ -406,6 +407,7 @@ export class ApolloServerBase {
             schemaHash: this.schemaHash,
             engine: {
               serviceID: this.engineServiceId,
+              apiKeyHash: this.engineApiKeyHash,
             },
             persistedQueries: this.requestOptions.persistedQueries,
           }),