+++ packages/apollo-server-core/src/graphqlOptions.ts
@@ -3,6 +3,8 @@ import {
   ValidationContext,
   GraphQLFieldResolver,
   DocumentNode,
+  GraphQLError,
+  GraphQLFormattedError,
 } from 'graphql';
 import { GraphQLExtension } from 'graphql-extensions';
 import { CacheControlExtensionOptions } from 'apollo-cache-control';
@@ -10,6 +12,8 @@ import { KeyValueCache, InMemoryLRUCache } from 'apollo-server-caching';
 import { DataSource } from 'apollo-datasource';
 import { ApolloServerPlugin } from 'apollo-server-plugin-base';
 import { GraphQLParseOptions } from 'graphql-tools';
+import { ValueOrPromise } from 'apollo-server-env';
+import { GraphQLExecutor } from '../dist/requestPipelineAPI';
 
 /*
  * GraphQLServerOptions
@@ -31,10 +35,11 @@ export interface GraphQLServerOptions<
   TRootValue = any
 > {
   schema: GraphQLSchema;
-  formatError?: Function;
+  formatError?: (error: GraphQLError) => GraphQLFormattedError;
   rootValue?: ((parsedQuery: DocumentNode) => TRootValue) | TRootValue;
   context?: TContext | (() => never);
   validationRules?: Array<(context: ValidationContext) => any>;
+  executor?: GraphQLExecutor;
   formatResponse?: Function;
   fieldResolver?: GraphQLFieldResolver<any, TContext>;
   debug?: boolean;
@@ -62,9 +67,7 @@ export default GraphQLServerOptions;
 export async function resolveGraphqlOptions(
   options:
     | GraphQLServerOptions
-    | ((
-        ...args: Array<any>
-      ) => Promise<GraphQLServerOptions> | GraphQLServerOptions),
+    | ((...args: Array<any>) => ValueOrPromise<GraphQLServerOptions>),
   ...args: Array<any>
 ): Promise<GraphQLServerOptions> {
   if (typeof options === 'function') {