+++ packages/apollo-server-core/src/types.ts
@@ -5,8 +5,11 @@ import {
   IMocks,
   GraphQLParseOptions,
 } from 'graphql-tools';
+import { ValueOrPromise } from 'apollo-server-env';
 import { ConnectionContext } from 'subscriptions-transport-ws';
-import WebSocket from 'ws';
+// The types for `ws` use `export = WebSocket`, so we'll use the
+// matching `import =` to bring in its sole export.
+import WebSocket = require('ws');
 import { GraphQLExtension } from 'graphql-extensions';
 export { GraphQLExtension } from 'graphql-extensions';
 
@@ -27,10 +30,10 @@ export { GraphQLSchemaModule };
 
 export { KeyValueCache } from 'apollo-server-caching';
 
-export type Context<T = any> = T;
-export type ContextFunction<T = any> = (
-  context: Context<T>,
-) => Promise<Context<T>>;
+export type Context<T = object> = T;
+export type ContextFunction<FunctionParams = any, ProducedContext = object> = (
+  context: FunctionParams,
+) => ValueOrPromise<Context<ProducedContext>>;
 
 // A plugin can return an interface that matches `ApolloServerPlugin`, or a
 // factory function that returns `ApolloServerPlugin`.
@@ -48,11 +51,12 @@ export interface SubscriptionServerOptions {
 }
 
 type BaseConfig = Pick<
-  GraphQLOptions<Context<any>>,
+  GraphQLOptions<Context>,
   | 'formatError'
   | 'debug'
   | 'rootValue'
   | 'validationRules'
+  | 'executor'
   | 'formatResponse'
   | 'fieldResolver'
   | 'tracing'
@@ -69,11 +73,11 @@ export interface Config extends BaseConfig {
   resolvers?: IResolvers;
   schema?: GraphQLSchema;
   schemaDirectives?: Record<string, typeof SchemaDirectiveVisitor>;
-  context?: Context<any> | ContextFunction<any>;
+  context?: Context | ContextFunction;
   introspection?: boolean;
   mocks?: boolean | IMocks;
   mockEntireSchema?: boolean;
-  engine?: boolean | EngineReportingOptions<Context<any>>;
+  engine?: boolean | EngineReportingOptions<Context>;
   extensions?: Array<() => GraphQLExtension>;
   cacheControl?: CacheControlExtensionOptions | boolean;
   plugins?: PluginDefinition[];