+++ packages/apollo-server-core/src/utils/isNode.ts
@@ -1,6 +1,0 @@
-export default typeof process === 'object' &&
-  process &&
-  process.release &&
-  process.release.name === 'node' &&
-  process.versions &&
-  typeof process.versions.node === 'string';